Client listing page and clients page refactored into one as they were duplicates, created AddClient component but not fleshed out
