+++ src/main/webapp/resources/js/apis/clients/clients.js
@@ -1,4 +1,10 @@
+/**
+ * @fileOverview This file is responsible for all asynchronous calls for
+ * clients.
+ */
+
 import axios from "axios";
+import { setBaseUrl } from "../../utilities/url-utilities";
 
 /**
  * Revoke all tokens for a given client by identifier.
@@ -7,4 +13,12 @@ import axios from "axios";
  */
 export async function revokeClientTokens(id) {
   return await axios.delete(`clients/ajax/revoke?id=${id}`);
+}
+
+/**
+ * Add new client.
+ * @return {Promise<AxiosResponse<T>>}
+ */
+export async function addNewClient() {
+  return await axios.post(`clients/create`);
 }
